AST Service
The AST service is to provide the user with the AST tree.

The current state of progress is that the endpoint exists, and the
result should be result in the form of a ServiceResult (which
AstServiceResult extends).

The AstServiceDispatcher will return this result, and this in turn will
return it to the user.